build: Regenerate uninames.h
To avoid integer overflow in the unicode_name_by_length table.

https://bugzilla.gnome.org/show_bug.cgi?id=771223

+++ gllib/uniname/uninames.h
@@ -11791,7 +11791,7 @@ static const char unicode_name_words[71408] = {
   'P', 'A', 'S', 'S', 'I', 'V', 'E', '-', 'P', 'U', 'L', 'L', '-', 'D', 'O', 'W', 'N', '-', 'O', 'U', 'T', 'P', 'U', 'T',
 };
 #define UNICODE_CHARNAME_NUM_WORDS 11782
-static const struct { uint16_t extra_offset; uint16_t ind_offset; } unicode_name_by_length[26] = {
+static const struct { uint32_t extra_offset; uint16_t ind_offset; } unicode_name_by_length[26] = {
   { 0, 0 },
   { 0, 0 },
   { 26, 26 },
